Initial version of the vim script!
It's rough right now; just want to get this in now since it's a good
stopping point. Some associated changes to the font are also included.
They are kind of awful -- some subscript numerals had to be duplicated
and some other miscellaneous characters had to be moved around.  Anyway,
it's all there.